alerts: WIP of working alerts analysis

+++ alerts/src/utils/pct.rs
@@ -13,8 +13,8 @@ pub fn compute_pct(results: &[PctDTORaw]) -> f64 {
     let (prev_div, curr_div) = (results[1].divisor, results[0].divisor);
     let (prev_num, curr_num) = (results[1].numerator, results[0].numerator);
     // Compute the delta value between both previous and current
-    let total_d = ((curr_div + curr_num) - (prev_div + prev_num)) as f64;
-    let divisor_d = (curr_div - prev_div) as f64;
+    let total_d = (curr_div + curr_num) - (prev_div + prev_num);
+    let divisor_d = curr_div - prev_div;
 
     // Return the computed percentage
     ((total_d - divisor_d) / total_d) * 100.0
